+++ plugins/org.eclipse.n4js/src/org/eclipse/n4js/validation/flowValidators/SymbolDeclaration.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.flowgraphs.analysers;
+package org.eclipse.n4js.validation.flowValidators;
 
 import org.eclipse.emf.ecore.EObject;
 import org.eclipse.n4js.flowgraphs.dataflow.symbols.Symbol;
